use regmap
While this means that temperature and temperature limits need
two operations, the benefit of caching by far outweighs the downside.
of having to execute two i2c commands to read the temperature.

Use regmap_bulk commands to read and write temperature and
temperature limits to avoid having to introduce a local lock.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>